fix: remove gulp
