+++ package.json
@@ -10,7 +10,7 @@
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "ghpages": "gulp prepareGhPages",
+    "ghpages": "node scripts/deploy.js",
     "predeploy": "blockly-scripts predeploy",
     "prepublishOnly": "npm login --registry https://wombat-dressing-room.appspot.com",
     "start": "blockly-scripts start",
@@ -64,7 +64,6 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-jsdoc": "^46.8.2",
     "globals": "^15.4.0",
-    "gulp": "^5.0.1",
     "html-webpack-plugin": "^5.6.0",
     "mocha": "^11.1.0",
     "p5": "^1.10.0",
